% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes2labels.R
\name{codes2labes}
\alias{codes2labes}
\title{Format data of \code{\link{istac_get}} return from codes to labels}
\usage{
codes2labes(datos_lista, df)
}
\arguments{
\item{datos_lista}{a list given by \code{\link{istac_get}}}

\item{df}{a data frame with data in Code format.}
}
\value{
a data frame with the column names changed accordingly
}
\description{
change the data of dataframe returned by \code{\link{istac_get}}
from the code gived by th ISTAC API to the labels gived them.
}
