% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istac.R
\name{istac}
\alias{istac}
\title{Download data from the ISTAC API}
\usage{
istac(istac_table, islas = "all", label = FALSE, startdate, enddate, freq,
  mrv, POSIXct = FALSE, cache)
}
\arguments{
\item{istac_table}{Character string with the ID code of the table requested.
This ID code corresponds to the \code{namecode} column from \code{\link{cache}}.}

\item{islas}{Character vector of islands codes requested. Default value is special code of \code{all}.}

\item{label}{if \code{FALSE}, the data frame returned has the codes used in ISTAC API,
if \code{TRUE}, the data frame returned has the labels used in ISTAC API. Default value is \code{FALSE}.}

\item{startdate}{Numeric or character. If numeric it must be in \%Y form (i.e. four digit year).
For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01"
and for quarterly data, "2016Q1". This also accepts a special value of "YTD", useful for more frequently
updated subannual indicators.}

\item{enddate}{Numeric or character. If numeric it must be in \%Y form (i.e. four digit year).
For data at the subannual granularity the API supports a format as follows: for monthly data, "2016M01"
and for quarterly data, "2016Q1".}

\item{mrv}{Numeric. The number of Most Recent Values to return. A replacement of \code{startdate} and \code{enddate},
this number represents the number of observations you which to return starting from the most recent date of collection.
Useful in conjuction with \code{freq}.}
}
\value{
Data frame with all available requested data.
}
\description{
This function downloads the requested information using the ISTAC API.
}
