% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istac_get.R
\name{istac_get}
\alias{istac_get}
\title{Call the ISTAC API and return a formatted data frame}
\usage{
istac_get(indicador)
}
\arguments{
\item{indicador}{A character string. The \code{ID} code of the requested table.
Normally gived by the \code{\link{istac_search}} function.}
}
\value{
A list with data in json format and a formatted data frame.
}
\description{
This function calls the ISTAC API, capture the data in json
format and return a list with data in json format and a formatted data frame.
}
\details{
# To be used inside of \code{\link{istac}}
}
\examples{
# Percentiles de renta disponible (año anterior al de la entrevista) por hogar en Canarias y años.
istac_get("soc.cal.enc.res.3637")
}
