% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istacbase.R
\name{istacbase}
\alias{istacbase}
\title{Download data from the ISTACBASE API}
\usage{
istacbase(
  istacbase_table,
  islas = "all",
  label = TRUE,
  POSIXct = FALSE,
  startdate,
  enddate,
  freq,
  mrv,
  cache
)
}
\arguments{
\item{istacbase_table}{Character string with the \code{ID} of the table requested.
This \code{ID} corresponds to the \code{ID} column from \code{\link{cache}}.}

\item{islas}{Character vector of islands name requested. Default value is special code of \code{all}.
Valid values for islands are: El Hierro, La Palma, La Gomera, Tenerife, Gran Canaria, Fuerteventura and Lanzarote.}

\item{label}{if \code{FALSE}, the data frame returned has the codes used in ISTACBASE API,
if \code{TRUE}, the data frame returned has the labels used in ISTACBASE API. Default value is \code{FALSE}.}

\item{POSIXct}{if \code{TRUE}, additonal columns \code{fecha} and \code{periodicidad} are added.
\code{fecha} converts the default date into a \code{\link[base]{POSIXct}}. \code{periodicidad}
denotes the time resolution that the date represents.  Useful for \code{freq} filter,  If \code{FALSE}, these fields are not added.}

\item{startdate}{Numeric. Must be in \%Y form (i.e. four digit year).}

\item{enddate}{Numeric. Must be in \%Y form (i.e. four digit year).}

\item{freq}{Character String. For fetching yearly ("anual"), biannual ("semestral"), quaterly ("trimestral"), monthly("mensual"), bi-weekly("quincenal"), weekly("semanal") values.
Currently works along with \code{mrv}.}

\item{mrv}{Numeric. The number of Most Recent Values to return. A replacement of \code{startdate} and \code{enddate},
this number represents the number of observations you which to return starting from the most recent date of collection.}

\item{cache}{Data frame with tables from ISTACBASE API.}
}
\value{
Data frame with all available requested data.
}
\description{
This function downloads the requested information using the ISTACBASE API.
}
\note{
The \code{POSIXct} parameter requries the use of \code{\link[lubridate]{lubridate}} (>= 1.5.0). All dates
 are rounded down to the floor. For example a value for the year 2016 would have a \code{POSIXct} date of
 \code{2016-01-01}. If this package is not available and the \code{POSIXct} parameter is set to \code{TRUE},
 the parameter is ignored and a \code{warning} is produced.

 \code{startdate}, \code{enddate}, \code{freq}, \code{mrv} with \code{POSISXct}=\code{FALSE} are ignored when POSCIXct is set to FALSE.
and a \code{warning} is produced.
}
\examples{
# Percentiles de renta disponible (año anterior al de la entrevista) por hogar en Canarias y años.
istacbase("soc.cal.enc.res.3637")

# query using startdate and enddate
# Percentiles de renta disponible (año anterior al de la entrevista) por hogar en Canarias y años.
istacbase("soc.cal.enc.res.3637", POSIXct = TRUE, startdate = 2010, enddate = 2015)


# query using \code{islas} filter
# Población según sexos y edades año a año. Islas de Canarias y años.
istacbase("dem.pob.exp.res.35", islas = "Fuerteventura")

# if you want the most recent values
istacbase("dem.pob.exp.res.35", POSIXct = TRUE, mrv = 4)



}
