% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istacbase_get.R
\name{istacbase_get}
\alias{istacbase_get}
\title{Call the ISTACBASE API and return a formatted data frame}
\usage{
istacbase_get(indicador)
}
\arguments{
\item{indicador}{A character string. The \code{ID} code of the requested table.
Normally gived by the \code{\link{istacbase_search}} function.}
}
\value{
A list with data in json format and a formatted data frame.
}
\description{
This function calls the ISTACBASE API, capture the data in json
format and return a list with data in json format and a formatted data frame.
}
\details{
# To be used inside of \code{\link{istacbase}}
}
\examples{
# Percentiles de renta disponible (año anterior al de la entrevista) por hogar en Canarias y años.
istacbase_get("soc.cal.enc.res.3637")
}
