% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istacperiodos2POSIXct.R
\name{istacperiodos2POSIXct}
\alias{istacperiodos2POSIXct}
\title{Add a POSIXct dates to a ISTAC API return}
\usage{
istacperiodos2POSIXct(df, date_col)
}
\arguments{
\item{df}{data frame returned from API call}

\item{date_col}{name of the current date field}
}
\value{
If the package lubridate (>= 1.5.0) is available the original data frame with two new columns,
\code{fecha} and \code{periodicidad} is returned. If the above package is not available,
the orignal data frame is returned unaltered with an additional warning message.
}
\description{
Add a POSIXct date column as well as a column with the
appropreiate granularity to a ISTAC API return
}
