% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istac_search.R
\name{istac_search}
\alias{istac_search}
\title{Search information about data available through ISTAC API}
\usage{
istac_search(pattern, fields = "title", extra = TRUE, cache)
}
\arguments{
\item{pattern}{Character string or regular expression to be matched.}

\item{fields}{Character vector of column names through which to search}

\item{extra}{if \code{FALSE}, only the ID and Title are returned,
if \code{TRUE}, all columns of the \code{cache} are returned.}

\item{cache}{Data frame with metadata about API and ISTAC information.}
}
\value{
Data frame with metadata that match the search term.
}
\description{
This funcion allows finds those ID that match the search term and
returns a data frame with results
}
\examples{
istac_search(pattern = "Superficie")
istac_search(pattern = "Superficie", extra = TRUE)

istac_search(pattern = "medio ambiente", fields = "Estadísticas por temas")
istac_search(pattern = "medio ambiente", fields = "Estadísticas por temas", extra = TRUE)

}
